adding ability to store data with task messages (start, end, startcomp, endcomp) (#15)
Adding the ability to store opaque data with Task Messages (StartTask, EndTask, StartCompTask, EndCompTask) this is useful for debugging / diagnostics. Its also useful for storing intermediary results which can be rolled up when a Saga has Completed, or generating reports after the fact from the SagaLog.

Exposing this data to the applications via the SagaState API, and can be added via Saga API for logging messages.