Saga Recovery (#27)
Saga Recovery will happen at Startup. Once a Scheduler starts up it should read, it's saga log, and then reschedule non completed work. Startup returns a list of SagaIds, RecoverSagaState should then be called per SagaId. These methods are separate to allow the caller to determine how to handle failures (how many times to retry etc...) also to handle the rate at which the SagaLog is polled, and how quickly tasks are rescheduled. These decisions are specific to an implementation and therefore should not be abstracted from the user by the Saga package.

This code review also contains Saga Recovery Logic, which reads Sagas from the SagaLog, replays messages and verifies state is correct, and checks if its safe to proceed otherwise aborts the Saga.

The demo has been updated to poll the InMemorySagaLog (which just returns the ids of all sagas started since creation) after running all the work. It then verifies that all sagas are in the completed state.