Property Based Testing for Sagas (#31)
I tested out property based testing in Go using gopter. Property based testing allows us to define properties that our code should have and then generate input. If a counterexample is found the test fails and the input that caused it to fail is returned. This is great because it helps us explore more of the state space of possible inputs into our program. I also personally like it because using propositional logic to define the properties of a system typically helps me uncover missing cases in my code!

While doing property based testing I found some missed cases which I have updated in the SagaState Update method. This is also in this code review.

I did it first for SagaState updateState method. This is the core part of the Saga State Machine logic, and has a multitude of inputs. Too many to enumerate in unit tests. These are defined in sagaState_prop_test.go and run as part of our normal test suite.

I added Saga Specific Generators for property based testing to sagaGenerators.go . These are methods that help us randomly generate a valid SagaState or Ids.

I also removed a bunch of unit tests from sagaState_test.go that are covered in the property based tests. The property based tests are more succinct and have the same code coverage.