snapshots: Checkouter; the first piece of Filer
Filer offers high-level interaction with Snapshots. Checkout a Snapshot into the
local filesystem, or Ingest the local filesystem into a Snapshot. (For now we
only offer Checkout).

Use this abstraction in Runner and extend the Execer interface to take a Dir.
This means that Runner knows about Scoot (it deals with Snapshots), but Execer
doesn't.

Restructure the run code in Runner. It's now more clear who's responsible for
modifying what. This was necessary because checking out may take some time,
and we want to let the user cancel the job while it's happening. This has the
slight negative that the Worker returning to the client no longer means the
command is running. It would be nice if we could reinstate this. In the common
case, we have to wait for a quick checkout, but we may have to perform a fairly
slow git fetch (or a large checkout). ScootFS will eventually solve this
(because it can fault in the data).

Next steps are:

Make a Checkouter that actually checks out (from git, e.g.)

Use this directory in os execer

Allow sim_execer to examine this data (for testing)