Add Async Library (#86)
Add an Async Library to Scoot which will facilitate receiving return values (errors only for now) from go functions. For an example of how this should be used see the godoc comments in mailbox.go. AsyncError is similar to a Promise that contains an error value.

This helps us run an event loop in the scheduler with go routines and then executing callbacks when the go routine completes (successfully or unsuccessfully).