Caitie/stateful sched (#93)
Adding a Stateful Scheduler. For now I've left the old stateless scheduler in as well to keep the review small, once this goes in I'll update config and other binaries in an additional pass and remove old files.

Stateful Scheduler keeps cluster state and job state so it can do a better job of scheduling jobs. Currently it just schedules them on anyone that is free, but now we have a structure that can evolve over time.

StatefulScheduler's main work is done in an event loop. Every step it processes any messages received using Async library, and then decides which tasks to schedule and kicks off async tasks to do so.