[scheduler] refactor - move jobId generation to scheduler (#109)
Currently the JobId is generated in ScootAPI.  Logically it makes more sense for the scheduler to own this and return an Id if the job was successfully scheduled.  No functionality changes occur in this code review

No functionality Changed
* Updated Scheduler Interface to take a JobDefinition and return a JobId 
* Updated Generators 
* Moved JobId generation from scootapi to statefulscheduler

Also i have an LGTM Its just being slow to run!