Caitie/max retry task (#152)
Adding a configurable value to the scheduler MaxRetryTasks. This value is the maximum number of times a task will be retried before being marked as completed. This prevents a bad task from taking down the entire cluster. Total times tried is 1 + MaxRetryTasks.

Once a Task is Retried maximum number of times and has still failed the Task is marked as completed and the error is recorded in the Saga. This allows the rest of the Job to progress and try to execute.
Add SchedulerConfig to DI/config system to ease the pain & reduce the amount of code changes necessary for adding config values in the future.
There is a separate SchedulerConfig object in the Scheduler package to avoid coupling our config implementation to the Scheduler implementation. This requires some extra code, but I think is the right separation of concerns in the future.
additional unit test cases added
verified by running swarmtest on a local cluster.