[scheduler] Caitie/sagalogconfig (#172)
we want to be able to configure the SagaLog to be in memory or potentially a durable option. Adding config options to make this easy to specify.

I also refactored scootapi/setup.go to explicitly specify all inputs and outputs into MagicBag like we did with worker server, also added all dependencies. So no additions are necessary but all options can be overridden.