scootapi/setup: improve waiting and killing (#182)
Problem: if you start a command with Cmds, and it then starts another command, Cmds.Kill() will not kill it. This is especially problematic if the command it starts is a server that binds to a port.

Solution: Cmds.Kill now sends the signal to the entire process group. It also starts processes in a new process group.

(We weren't hitting this in open source, but were internally, where our saga log implementation is a server on a port)