retry smoke test & dial update (#188)
Currently SmokeTest fails as soon as the scheduler dies. Any call to ScootAPI that fails results in the test failing. However, we want to test scheduler restarts, which means some calls may fail.

This update allows failures to the scheduler to be retried. Currently by reusing the ScootApiClient associated with the Dialer, a broken pipe will be broken forever so Dial() needs to create a new scootClient connection on every call