saga: copy on get and logMessage, not update.  (#199)
Getting the status for a Job should be quick (even a second is longer than ideal). It currently takes several minutes for a job with 8000 tasks. Eep!

Getting the status requires reading the Saga State. Constructing the Saga State requires Recovery.

Saga recovery has to replay each message (which is O(number of messages)). Currently, updateSagaState copies the Saga State (which is O(size of state)).

This means that the time to recover a saga is superlinear (always copying a growing object).

This isn't needed on recovery, because we trust the messages are valid. Move the copy out of recovery and do it only on get (when the state is handed to a client) and in logMessage (where the write to durable storage may fail, and then we'll want to revert the state). This makes updateSagaState mutate, and reduces the cost of saga recovery from minutes (and growing) to seconds (even up to 15,000 tasks).