sched/scheduler: update status of tasks when they start running (#222)
We want to be able to see the status of tasks as they run (so we can, e.g., follow the output as it goes)

This requires using runner.Service directly in task_runner.go

Thus, we can get rid of Worker. Yay.

This change got big because it requires config changes.